requirements.txt
asgiref==3.7.2
astroid==3.0.2
attrs==23.2.0
black==23.12.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.7
Django
django-debug-toolbar==4.2.0
django-filter==23.5
django-phone-field==1.8.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
flake8==6.1.0
idna==3.6
inflection==0.5.1
isort==5.13.2
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
Markdown==3.5.1
MarkupSafe==2.1.3
mccabe==0.7.0
mypy-extensions==1.0.0
openapi-codec==1.3.2
packaging==23.2
pathspec==0.12.1
platformdirs==4.1.0
psycopg2-binary
pycodestyle==2.11.1
pyflakes==3.1.0
PyJWT==2.8.0
pylint==3.0.3
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.16.2
simplejson==3.19.2
sqlparse==0.4.4
tomlkit==0.12.3
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
